+++ default.nix
@@ -69,41 +69,93 @@ rec {
 
   # Have rustc spit out unstable target config json so we can do a minimum of
   # hard-coding.
-  stockThumbTarget = pkgs.runCommand "stock-target.json" {
-    nativeBuildInputs = [ pkgs.buildPackages.rustcRopi ];
-  } ''
-    rustc -Z unstable-options --print target-spec-json --target thumbv6m-none-eabi > $out
-  '';
+  stockThumbTargets = lib.genAttrs
+    [ "thumbv6m-none-eabi" "thumbv8m.main-none-eabi" ]
+    (target-name:
+      pkgs.runCommand "stock-target.json" {
+        nativeBuildInputs = [ pkgs.buildPackages.rustcRopi ];
+      } ''
+        rustc -Z unstable-options --print target-spec-json --target ${target-name} > $out
+      '');
 
-  ledgerPkgs = pkgsFunc {
-    config.allowUnsupportedSystem = true;
-    crossSystem = {
+  perDevice = let
+    f = crossSystem: import ./per-device.nix {
+      inherit
+        pkgs
+        rustPlatform
+        cargo-ledger
+        cargo-watch
+        ledgerctl
+        speculos
+        stack-sizes
+        ;
+      ledgerPkgs = pkgsFunc {
+        config.allowUnsupportedSystem = true;
+        inherit crossSystem;
+        inherit localSystem overlays;
+      };
+    };
+  in {
+    nanos = f {
       isStatic = true;
       config = "armv6m-unknown-none-eabi";
+      libc = "newlib-nano";
       gcc = {
         arch = "armv6s-m";
       };
-      rustc = {
+      rustc = rec {
         config = "thumbv6m-none-eabi";
-        platform = builtins.fromJSON (builtins.readFile stockThumbTarget) // {
+        platform = builtins.fromJSON (builtins.readFile stockThumbTargets.${config}) // {
           is-builtin = false;
 
-          # Shoudn't be needed, but what rustc prints out by default is
-          # evidently wrong!
-          atomic-cas = true;
+          atomic-cas = false;
           os = "nanos";
-          target-family = "bolos";
+          target-family = [ "bolos" ];
+
+          relocation-model = "ropi";
         };
       };
     };
-    inherit localSystem overlays;
-    crossOverlays = [
-      (self: super: {
-        newlibCross = super.newlibCross.override {
-          nanoizeNewlib = true;
+    nanox = f {
+      isStatic = true;
+      config = "armv6m-unknown-none-eabi";
+      libc = "newlib-nano";
+      gcc = {
+        arch = "armv6s-m";
+      };
+      rustc = rec {
+        config = "thumbv6m-none-eabi";
+        platform = builtins.removeAttrs (builtins.fromJSON (builtins.readFile stockThumbTargets.${config})) ["features"] // {
+          is-builtin = false;
+
+          atomic-cas = false;
+          os = "nanox";
+          target-family = [ "bolos" ];
+
+          relocation-model = "ropi-rwpi";
         };
-      })
-    ];
+      };
+    };
+    nanosplus = f {
+      isStatic = true;
+      config = "armv8m-unknown-none-eabi";
+      libc = "newlib-nano";
+      gcc = {
+        arch = "armv8-m.main";
+      };
+      rustc = rec {
+        config = "thumbv8m.main-none-eabi";
+        platform = builtins.fromJSON (builtins.readFile stockThumbTargets.${config}) // {
+          is-builtin = false;
+
+          max-atomic-width = 32;
+          os = "nanosplus";
+          target-family = [ "bolos" ];
+
+          relocation-model = "ropi-rwpi";
+        };
+      };
+    };
   };
 
   # TODO: Replace this with `thunkSource` from nix-thunk for added safety
@@ -125,79 +177,6 @@ rec {
 
   crate2nix = import ./dep/crate2nix { inherit pkgs; };
 
-  buildRustPackageClang = ledgerRustPlatform.buildRustPackage.override {
-    stdenv = ledgerPkgs.lldClangStdenv;
-  };
-
-  # TODO once we break up GCC to separate compiler vs runtime like we do with
-  # Clang, we shouldn't need these hacks to get make the gcc runtime available.
-  gccLibsPreHook = ''
-    export NIX_LDFLAGS
-    NIX_LDFLAGS+=' -L${ledgerPkgs.stdenv.cc.cc}/lib/gcc/${ledgerPkgs.stdenv.hostPlatform.config}/${ledgerPkgs.stdenv.cc.cc.version}'
-  '';
-
-  # Our tools are named differently than the Cargo defaults.
-  cargoLedgerPreHook = ''
-    export CARGO_TARGET_THUMBV6M_NONE_EABI_OBJCOPY=$OBJCOPY
-    export CARGO_TARGET_THUMBV6M_NONE_EABI_SIZE=$SIZE
-  '';
-
-  rustShell = buildRustPackageClang {
-    stdenv = ledgerPkgs.lldClangStdenv;
-    name = "rustShell";
-    src = null;
-    # We are just (ab)using buildRustPackage for a shell. When we actually build
-    __internal_dontAddSysroot = true;
-    preHook = gccLibsPreHook;
-    shellHook = cargoLedgerPreHook;
-    # We just want dev shell
-    unpackPhase = ''
-      echo got in shell > $out
-      exit 0;
-    '';
-    cargoVendorDir = "pretend-exists";
-    depsBuildBuild = [ ledgerPkgs.buildPackages.stdenv.cc ];
-    inherit (ledgerPkgs.alamguRustPackages.rustPlatform) rustLibSrc;
-    nativeBuildInputs = [
-      # emu
-      speculos.speculos ledgerPkgs.buildPackages.gdb
-
-      # loading on real hardware
-      cargo-ledger
-      ledgerctl
-
-      # just plain useful for rust dev
-      stack-sizes
-      cargo-watch
-
-      # Testing stuff against nodejs modules
-      pkgs.nodejs_latest
-    ];
-    verifyCargoDeps = true;
-    target = "thumbv6m-none-eabi";
-
-    # Cargo hash must be updated when Cargo.lock file changes.
-    cargoSha261 = "1kdg77ijbq0y1cwrivsrnb9mm4y5vlj7hxn39fq1dqlrppr6fdrr";
-
-    # It is more reliable to trick a stable rustc into doing unstable features
-    # than use an unstable nightly rustc. Just because we want unstable
-    # langauge features doesn't mean we want a less tested implementation!
-    RUSTC_BOOTSTRAP = 1;
-
-    meta = {
-      platforms = lib.platforms.all;
-    };
-  };
-
-  # Use right Rust; use Clang.
-  buildRustCrateForPkgsLedger = pkgs: let
-    isLedger = (pkgs.stdenv.hostPlatform.rustc.platform.os or "") == "nanos";
-    platform = if isLedger then ledgerRustPlatform else rustPlatform;
-  in pkgs.buildRustCrate.override rec {
-    stdenv = if isLedger then pkgs.lldClangStdenv else pkgs.stdenv;
-    inherit (platform.rust) rustc cargo;
-  };
-
   rustPlatform = pkgs.makeRustPlatform {
     inherit (pkgs.alamguRustPackages) cargo;
     # Go back one stage too far back (`buildPackages.buildPackages` not
@@ -205,13 +184,6 @@ rec {
     # stdlib from scratch we don't need a "cross compiler" --- rustc itself is
     # actually always multi-target.
     rustc = pkgs.buildPackages.buildPackages.rustcRopi;
-  };
-
-  ledgerRustPlatform = ledgerPkgs.makeRustPlatform {
-    inherit (pkgs.alamguRustPackages) cargo;
-    rustcSrc = ledgerPkgs.buildPackages.rustcBuilt.src;
-    # See above for why `buildPackages` twice.
-    rustc = ledgerPkgs.buildPackages.buildPackages.rustcRopi;
   };
 
   ledgerctl = with pkgs.python3Packages; buildPythonPackage {
@@ -232,60 +204,9 @@ rec {
     ];
   };
 
-  buildRustCrateForPkgsWrapper = pkgs: fun: let
-    isLedger = (pkgs.stdenv.hostPlatform.rustc.platform.os or "") == "nanos";
-  in args: fun (args // lib.optionalAttrs isLedger {
-      RUSTC_BOOTSTRAP = true;
-      extraRustcOpts = [
-        "-C" "relocation-model=ropi"
-        "-C" "passes=ledger-ropi"
-        "-C" "opt-level=3"
-        "-C" "codegen-units=1"
-        "-C" "embed-bitcode"
-        "-C" "lto"
-        "-Z" "emit-stack-sizes"
-        # Otherwise we don't run our custom pass
-        "-Z" "new-llvm-pass-manager=no"
-        "--emit=link,dep-info,obj"
-      ] ++ args.extraRustcOpts or [];
-      # separateDebugInfo = true;
-      dontStrip = isLedger;
-    });
-
   generic-cli = (import ./node/cli {
     inherit pkgs;
   }).package;
-
-  ledgerStdlib = import ./stdlib/Cargo.nix {
-    pkgs = ledgerPkgs;
-    buildRustCrateForPkgs = pkgs: buildRustCrateForPkgsWrapper
-      pkgs
-      ((buildRustCrateForPkgsLedger pkgs).override {
-        defaultCrateOverrides = pkgs.defaultCrateOverrides // {
-          core = attrs: {
-            src = ledgerPkgs.alamguRustPackages.rustPlatform.rustLibSrc + "/core";
-            postUnpack = ''
-              cp -r ${ledgerPkgs.alamguRustPackages.rustPlatform.rustLibSrc}/stdarch $sourceRoot/..
-              cp -r ${ledgerPkgs.alamguRustPackages.rustPlatform.rustLibSrc}/portable-simd $sourceRoot/..
-            '';
-          };
-          alloc = attrs: { src = ledgerPkgs.alamguRustPackages.rustPlatform.rustLibSrc + "/alloc"; };
-          rustc-std-workspace-core = attrs: { src = ledgerPkgs.alamguRustPackages.rustPlatform.rustLibSrc + "/rustc-std-workspace-core"; };
-        };
-      });
-  };
-
-  ledgerCompilerBuiltins = lib.findFirst
-    (p: lib.hasPrefix "rust_compiler_builtins" p.name)
-    (builtins.throw "no compiler_builtins!")
-    ledgerStdlib.rootCrate.build.dependencies;
-
-  ledgerCore = lib.findFirst
-    (p: lib.hasPrefix "rust_core" p.name)
-    (builtins.throw "no core!")
-    ledgerStdlib.rootCrate.build.dependencies;
-
-  ledgerStdlibCI = ledgerStdlib.rootCrate.build;
 
   utils = import ./utils/Cargo.nix { inherit pkgs; };
 
@@ -294,4 +215,21 @@ rec {
   cargo-watch = utils.workspaceMembers.cargo-watch.build;
 
   stack-sizes = utils.workspaceMembers.stack-sizes.build;
+
+  # COMPAT
+  stockThumbTarget = stockThumbTargets.thumbv6m-none-eabi;
+  inherit (perDevice.nanos)
+    buildRustCrateForPkgsLedger
+    buildRustCrateForPkgsWrapper
+    buildRustPackageClang
+    cargoLedgerPreHook
+    gccLibsPreHook
+    ledgerCompilerBuiltins
+    ledgerCore
+    ledgerPkgs
+    ledgerRustPlatform
+    ledgerStdlib
+    ledgerStdlibCI
+    rustShell
+    ;
 }