+++ release.nix
@@ -46,4 +46,7 @@ in {
 }
   # Hack until CI will traverse contents
   // lib.mapAttrs' (n: lib.nameValuePair ("linux--" + n)) x86_64-linux
+  // lib.mapAttrs' (n: lib.nameValuePair ("linux--nanos--" + n)) x86_64-linux.perDevice.nanos
+  // lib.mapAttrs' (n: lib.nameValuePair ("linux--nanox--" + n)) x86_64-linux.perDevice.nanox
+  // lib.mapAttrs' (n: lib.nameValuePair ("linux--nanosplus--" + n)) x86_64-linux.perDevice.nanosplus
   // lib.mapAttrs' (n: lib.nameValuePair ("macos--" + n)) x86_64-darwin